+++ lua/continuous-testing/utils/file.lua
@@ -1,26 +1,7 @@
-local M = {}
-
-M.inject_file_to_test_command = function(test_command, file)
-    return test_command:gsub("%%file", file)
-end
-
-M.deepcopy_table = function(orig)
-    local orig_type = type(orig)
-    local copy
-    if orig_type == "table" then
-        copy = {}
-        for orig_key, orig_value in next, orig, nil do
-            copy[M.deepcopy_table(orig_key)] = M.deepcopy_table(orig_value)
-        end
-        setmetatable(copy, M.deepcopy_table(getmetatable(orig)))
-    else -- number, string, boolean, etc
-        copy = orig
-    end
-    return copy
-end
-
 local uv = vim.loop
 local validate = vim.validate
+
+local M = {}
 
 M.path = (function()
     local is_windows = uv.os_uname().version:match("Windows")
@@ -210,4 +191,5 @@ function M.find_first_ancestor(startpath, pattern)
         end
     end)
 end
+
 return M
